refactor: use types file over *.d.ts
