+++ client/src/utils/auth/types.ts
@@ -1,3 +1,5 @@
+import { NavigateFunction } from "react-router-dom";
+
 export interface FirebaseUtilParams {
   email: string;
   password: string;
